firstinit.sh - create link - bug fixed

+++ firstinit.sh
@@ -5,8 +5,8 @@ linkto()
     if [ -f "$1" -o -L "$1" ] && [ -z "$(grep -P "\\Q$2\\E\$" <(ls -l "$1") )" ]; then
 	mv "$1" "$1".bak-"$(date "+%F-%H-%M-%S")"
 	echo "$!" moved to "$1".bak-"$(date "+%F-%H-%M-%S")"
-	ln -s "$2" "$1"
     fi
+    [ -f "$1" -o -L "$1" ] || ln -s "$2" "$1"
 }
 
 linkto ~/.emacs ~/configs/_emacs.el
